refactor: remove /components/forms/pipeline folder
