+++ src/components/forms/index.ts
@@ -1,2 +1,1 @@
-export * from "./pipeline";
 export * from "./commons";
