--- src/components/forms/pipeline/ConfigurePipelineForm/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ConfigurePipelineForm";
-export * from "./ConfigurePipelineForm";
